Finalizar

+++ final/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 class Product(models.Model):
     name = models.CharField(max_length=50)
-    value = models.DecimalField(max_digits=8, decimal_places=2)
+    value = models.IntegerField()
 
     def __str__(self):
         return self.name
